requirements.txt
Flask>=2.3.0
python-dotenv>=1.0.0
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
Pillow>=10.0.0
flask-cors>=4.0.0
psycopg2-binary>=2.9.0
numpy>=1.24.0
gunicorn>=21.2.0
tensorflow>=2.13.0
opencv-python